DataTable new rows can have auto height. (#3213)
* DataTable new rows can have auto height.

Related issue: #3122.

* Test auto height computation in DataTable.add_row

* Add snapshot test for add_row height=None.

* Extract some styles logic into auxiliary methods.

When adding a row with automatic height, I need to render the cells to compute their height. Instead of wasting that rendering, I want to do it well and then cache it, which means I need to reuse some of the logic of the other rendering methods. By extracting some logic, I'll be able to reuse it.

* Cache auxiliary cell renderings.

* Fix test import.

* Set row height to 0 when adding auto-height row.

* Remove superfluous cache clear.

* Fix cache/typing issue.

* Cache method to compute styles to render cell.

We extract this logic into a method for two reasons.
For one, having this as a method with an lru cache enables caching these auxiliary styles, which don't depend directly on the location of the cell, but instead depend on the values of 9 Boolean flags (making for a total of 512 possible combinations, versus the infinite number of different positions/states a cell can be in.
Secondly, having this as a method allows me to compute these styles more easily from within _update_dimensions when trying to salvage the renderings of the cells of a new row that may have been pre-rendered with the wrong height.
(See the following commits for more context.)

* Perform surgery on the datatable cache.

* Improve data table tests.

* Reduce cache size.

The first five parameters (is_header_cell, is_row_label_cell, is_fixed_style_cell, hover, and cursor) are the ones that change more frequently, so it is reasonable to fix the size of the cache at 32.

Related comment: https://github.com/Textualize/textual/pull/3213#discussion_r1326071862

* Clear cache with other caches.

Related comment: https://github.com/Textualize/textual/pull/3213#discussion_r1326071862.