Fix snapshot test.
Now that text opacity is taken into account in the calculation of rich_style, the text color of disabled tabs became too light. To compensate for it, the text opacity in Tab:disabled could be set to 71% and that would match the previous value. However, 71% feels like a terrible value to have in CSS so I opted to go with 70% (which looks ALMOST EXACTLY THE SAME) and then I updated the snapshot test.