Switch to chunking Windows terminal sequences based on buffer size
First we tested this by splitting the output by lines; that worked but it
could be faster. So now we're going to take the problematic buffer
size (32KiB), assume a reasonable worst case of every character heading out
being 3 bytes in size, and chunking the output based on that.

In other words, experiment for #2548 take two.