Swap command palette isolation to a system-screen filtering approach
While for now, and likely for a while to come, the command palette is going
to be a special case we'll want to ignore, it's been decided that we should
introduce the concept of a system screen type and filter those out. So here
I introduce a _SystemModalScreen type and have the command palette inherit
from that, and then I filter that type out when looking for children of the
app.

Technically, at some point, I guess we should also have a _SystemScreen,
with a similar approach.

<whisper>This would be an excellent moment for a mixin</whisper>