Experiment to see if a call to close is needed too
While things are generally working fine on macOS (and possibly GNU/Linux,
that's still to be tested), there is the "can't input anything, have to kill
the terminal" issue on Windows. This worked in the PR a year ago, and this
bit of code seems to be the difference so let's test that out.