Ensure we don't restart application mode in the wrong place
It's possible for the developer to have code that is something like:

    with self.suspend():
        # do something here that we can Ctrl+Z.

such that the suspended process is *this* process; because of the signal
handlers involved, in this case, we wouldn't want to automatically restart
application mode. So this commit adds the ability to mark a body of code as
one where no auto-restart should take place.