docs: fix tcss syntax highlighting (#4176)
Sorry but I spotted that the syntax highlighting was wrong in my last PR to
add tcss to the docs (#4171) where I had blindly copy pasted the code.

A quick grep found a few other places in the docs where the syntax
highlighting was wrong. This PR corrects the language to css.