Only restore widget focus if the widget belongs to the current screen
While the terminal window didn't have focus, anything can could happen. The
widget could be removed, the screen could change, etc. So by the time
AppFocus happens the widget might not be one to focus any more.

Initially I was just making it the focused widget anyway and letting the
focus-handling code do what it needed to do. Sending focus to a widget that
isn't part of the DOM any more isn't exactly a breaking problem; but...

One issue is that you can end up with App.focused saying that a widget is
focused that isn't in the DOM any more. We don't want that. So here I'm a
bit more defensive. This changes things so that we check that the widget's
screen is still the screen that's in play. If the widget has been removed it
won't have a parent and so can't find its screen. All of this means that if
the screen has changed *or* if the widget has been removed, we're covered.