Don't restore focus on AppFocus if something has focus
While the application is in an AppBlur state, it's possible that some code
could have been running that updated what's focused. It doesn't make sense
to have Textual itself override the dev's choice to have focus be somewhere
else (perhaps the result of some long-running background process, that
they've tabbed away from, and when they tab back they expect to be in a
specific control).

So here I tweak the code that restores the focused widget so that it only
restores if it's still the case that nothing has focus.