Ensure scrollbars get Show/Hide messages
The idea here being that while scrollbars are attached to widgets, they
exist outwith of the DOM, and so don't seem to take part in the usual flow
of show/hide messaging. This change checks, when posting a show/hide
message, if the widget receiving the message has scrollbars and if it does
the message is also sent to them too.

Fixes #4274.