fix(datatable): change max-height back to 100%
The `max-height` of the `DataTable` was changed to 100vh rather than
100% in #3566, because at the time this caused issues with auto height
containers, as described in #2975.

However this issue was later fixed in #3814. This PR changes the
`max-height` of the `DataTable` back to 100%, as 100vh will actually
break common layouts and seems no longer required to workaround this
auto height issue.

Fixes #4286.