feat(highlight): use theme text colors
The new `highlight` module for syntax highlighting currently uses the
'base' theme colors. Perhaps this was intentional, but to my eye this
makes the colors in `Markdown` code blocks so prominent that they feel
a bit out of place.

This updates the highlight theme to instead use the dedicated theme text
colors, which I think looks more consistent.