test(modal): increase pauses in flaky test (#6116)
* test(modal): increase pauses in flaky test

It looks like `test_modal_pop_screen()` has started being flaky in CI.

Increase the initial pause to ensure the footer is fully composed when
clicked. This seemed to be the fix for a similar flaky test in #6032.

* test(modal): wait refresh for flaky test

After increasing the pause, this test is still flaky in CI. Try adding a
`wait_for_refresh` as well, the same as the fix for a similar flaky test
in #6116