fix(text area): fix cursor display on wrapped line
Currently the `TextArea` cursor display is broken when trying to
navigate a wrapped line:

- Clicking a wrapped line will cause the cursor to disappear
- Keyboard navigation on a wrapped line won't blink the cursor

The problem is that the line cache introduced in accc6da doesn't
properly account for wrapped lines.

Fix the cache key so the 'cursor location' element is not simply based
on a location in document-space.

Fixes #6195